Added namespaces which required a lot of other chagnes, see below.
Reworked the directory structure as well as renamed some files/classes.
Switched out the password hashing script for password_compat.
Kept PHPass, but only for the random string generator, used for
creating the session token.
Updated the cleanGit.sh script to account for the changes.
A lot more changes than I would typically like in one commit, but
in order to get things working, this was required.