Added Router class for custom routes configured via Routes.ini.
Altered Config class to read ini files. Updated bootstrap to
instantiate Router.

Changed $url variable in bootstrap, which was previously stored
in main.php, to be $uri. Also updated all instances of this
variable including where it is set in .htaccess.

Updated cleanGit.sh to account for php and ini files.

Updated example 'fruit' controller and model (to pull in
namespaced classes).